// frontend/src/app/core/generated/circabc/model/groupDeletionReport.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Node } from './node';
import { Profile } from './profile';


export interface GroupDeletionReport { 
    lockedNodes?: Array<Node>;
    sharedNodes?: Array<Node>;
    sharedProfiles?: Array<Profile>;
}